CHANGE split the variables class in multiple files
