+++ .gitignore
@@ -14,3 +14,4 @@
 /JoinFS/FS2020/Setup-FS2020/Release
 /JoinFS/FS2020/Setup-FS2020/Debug
 /JoinFS/P3D/bin
+/JoinFS/FSX/bin
